DummyLoader: move it out of core
This loader is supposed to not perform anything locally, and it's
actually a hack to bypass the loader when tests are to be resolved and
run remotely.

Its only users is the remote runner.  To avoid pollution in core,
let's move it to that module.
